Fix bugs for framework
