+++ phpunit.xml
@@ -20,7 +20,7 @@
             <directory suffix="Test.php">./src/defer/test</directory>
             <directory suffix="Test.php">./src/devtool/test</directory>
             <directory suffix="Test.php">./src/error-handler/test</directory>
-            <!--<directory suffix="Test.php">./src/framework/test</directory>-->
+            <directory suffix="Test.php">./src/framework/test</directory>
             <!--<directory suffix="Test.php">./src/http-client/test</directory>-->
             <directory suffix="Test.php">./src/http-server/test</directory>
             <!--<directory suffix="Test.php">./src/i18n/test</directory>-->
