+++ src/framework/test/Cases/AbstractTestCase.php
@@ -8,7 +8,7 @@
  * @license  https://github.com/swoft-cloud/swoft/blob/master/LICENSE
  */
 
-namespace SwoftTest;
+namespace SwoftTest\Cases;
 
 use PHPUnit\Framework\TestCase;
