+++ src/framework/test/Cases/BeanTest.php
@@ -7,13 +7,12 @@
  * @contact  group@swoft.org
  * @license  https://github.com/swoft-cloud/swoft/blob/master/LICENSE
  */
-namespace SwoftTest;
+namespace SwoftTest\Cases;
 
 use Swoft\App;
 use Swoft\Bean\Resource\ServerAnnotationResource;
-use Swoft\Proxy\Proxy;
-use SwoftTest\Bean\ProxyTest;
-use SwoftTest\Bean\TestHandler;
+use SwoftTest\Testing\Bean\ProxyTest;
+use SwoftTest\Testing\Bean\TestHandler;
 use SwoftTest\Testing\Bean\Config;
 
 /**
@@ -23,32 +22,13 @@ use SwoftTest\Testing\Bean\Config;
  */
 class BeanTest extends AbstractTestCase
 {
-    public function testProxy()
-    {
-        $object = new ProxyTest(1, 2);
-        $handler = new TestHandler($object);
-
-        /* @var ProxyTest $proxy */
-        $proxy = Proxy::newProxyInstance(ProxyTest::class, $handler);
-        $this->assertEquals('p11beforeafter', $proxy->publicParams('p1', 'p2'));
-        $this->assertEquals('p1p20beforeafter', $proxy->publicFun1('p1', 'p2'));
-        $this->assertEquals('p1p23beforeafter', $proxy->publicFun1('p1', 'p2', [1, 2, 3]));
-        $this->assertEquals('p1p21beforeafter', $proxy->publicFun2('p1', 'p2'));
-        $this->assertEquals('p1p21beforeafter', $proxy->publicFun3('p1', 'p2'));
-        $this->assertEquals('p1p25beforeafter', $proxy->publicFun3('p1', 'p2', 5));
-        $this->assertEquals('p1p2beforeafter', $proxy->publicFun1Base('p1', 'p2'));
-        $this->assertEquals('p1p21.6beforeafter', $proxy->publicFun2Base('p1', 'p2', 1.6));
-        $this->assertEquals('p1p2beforeafter', $proxy->publicFun1Trait('p1', 'p2'));
-        $this->assertEquals('p1p2beforeafter', $proxy->publicFun2Trait('p1', 'p2'));
-        $this->assertEquals('p1p2beforeafter', $proxy->publicFun3Trait('p1', 'p2'));
-    }
-
     public function testCustomComponentNamespaces()
     {
         $config = App::getProperties()->toArray();
         $this->assertArrayHasKey('components', $config);
         $resource = new ServerAnnotationResource($config);
         $resource->registerNamespace();
+        $resource->registerCustomNamespace();
 
         $namespace = $resource->getComponentNamespaces();
         $this->assertNotFalse(array_search('SwoftTest', $namespace));
