+++ src/framework/test/Testing/Pool/ConsulPptConfig.php
@@ -7,7 +7,7 @@
  * @contact  group@swoft.org
  * @license  https://github.com/swoft-cloud/swoft/blob/master/LICENSE
  */
-namespace SwoftTest\Pool;
+namespace SwoftTest\Testing\Pool;
 
 use Swoft\Bean\Annotation\Bean;
 use Swoft\Bean\Annotation\Value;
@@ -17,12 +17,12 @@ use Swoft\Bean\Annotation\Value;
  *
  * @Bean
  */
-class ConsulEnvConfig
+class ConsulPptConfig
 {
     /**
      * adress
      *
-     * @Value(env="${PROVIDER_CONSUL_ADDRESS}")
+     * @Value(name="${config.provider.consul.address}")
      * @var string
      */
     protected $address = 'http://127.0.0.1:80';
@@ -30,7 +30,7 @@ class ConsulEnvConfig
     /**
      * the tags of register service
      *
-     * @Value(env="${PROVIDER_CONSUL_TAGS}")
+     * @Value(name="${config.provider.consul.tags}")
      * @var array
      */
     protected $tags = [];
@@ -38,7 +38,7 @@ class ConsulEnvConfig
     /**
      * the timeout of consul
      *
-     * @Value(env="${PROVIDER_CONSUL_TIMEOUT}")
+     * @Value(name="${config.provider.consul.timeout}")
      * @var int
      */
     protected $timeout = 300;
@@ -46,7 +46,7 @@ class ConsulEnvConfig
     /**
      * the interval of register service
      *
-     * @Value(env="${PROVIDER_CONSUL_INTERVAL}")
+     * @Value(name="${config.provider.consul.interval}")
      * @var int
      */
     protected $interval = 3;
