+++ src/framework/test/Testing/Pool/DemoPoolConfig.php
@@ -1,5 +1,5 @@
 <?php
-namespace SwoftTest\Pool;
+namespace SwoftTest\Testing\Pool;
 
 use Swoft\Bean\Annotation\Bean;
 use Swoft\Bean\Annotation\Value;
