+++ src/framework/test/Testing/Pool/EnvAndPptPoolConfig.php
@@ -7,23 +7,23 @@
  * @contact  group@swoft.org
  * @license  https://github.com/swoft-cloud/swoft/blob/master/LICENSE
  */
-namespace SwoftTest\Pool;
+namespace SwoftTest\Testing\Pool;
 
 use Swoft\Bean\Annotation\Bean;
 use Swoft\Bean\Annotation\Value;
 use Swoft\Pool\PoolProperties;
 
 /**
- * config by properties
+ * properties and env
  *
  * @Bean
  */
-class PropertyPoolConfig extends PoolProperties
+class EnvAndPptPoolConfig extends PoolProperties
 {
     /**
      * the name of pool
      *
-     * @Value(name="${config.test.test.name}")
+     * @Value(name="${config.test.test.name}", env="${TEST_NAME}")
      * @var string
      */
     protected $name = '';
@@ -31,7 +31,7 @@ class PropertyPoolConfig extends PoolProperties
     /**
      * the maximum number of idle connections
      *
-     * @Value(name="${config.test.test.maxIdel}")
+     * @Value(name="${config.test.test.maxIdel}", env="${TEST_MAX_IDEL}")
      * @var int
      */
     protected $maxIdel = 6;
@@ -39,7 +39,7 @@ class PropertyPoolConfig extends PoolProperties
     /**
      * the maximum number of active connections
      *
-     * @Value(name="${config.test.test.maxActive}")
+     * @Value(name="${config.test.test.maxActive}", env="${TEST_MAX_ACTIVE}")
      * @var int
      */
     protected $maxActive = 50;
@@ -47,7 +47,7 @@ class PropertyPoolConfig extends PoolProperties
     /**
      * the maximum number of wait connections
      *
-     * @Value(name="${config.test.test.maxWait}")
+     * @Value(name="${config.test.test.maxWait}", env="${TEST_MAX_WAIT}")
      * @var int
      */
     protected $maxWait = 100;
@@ -55,7 +55,7 @@ class PropertyPoolConfig extends PoolProperties
     /**
      * the time of connect timeout
      *
-     * @Value(name="${config.test.test.timeout}")
+     * @Value(name="${config.test.test.timeout}", env="${TEST_TIMEOUT}")
      * @var int
      */
     protected $timeout = 200;
@@ -70,7 +70,7 @@ class PropertyPoolConfig extends PoolProperties
      * ]
      * </pre>
      *
-     * @Value(name="${config.test.test.uri}")
+     * @Value(name="${config.test.test.uri}", env="${TEST_URI}")
      * @var array
      */
     protected $uri = [];
@@ -78,7 +78,7 @@ class PropertyPoolConfig extends PoolProperties
     /**
      * whether to user provider(consul/etcd/zookeeper)
      *
-     * @Value(name="${config.test.test.useProvider}")
+     * @Value(name="${config.test.test.useProvider}", env="${TEST_USE_PROVIDER}")
      * @var bool
      */
     protected $useProvider = false;
@@ -86,7 +86,7 @@ class PropertyPoolConfig extends PoolProperties
     /**
      * the default balancer is random balancer
      *
-     * @Value(name="${config.test.test.balancer}")
+     * @Value(name="${config.test.test.balancer}", env="${TEST_BALANCER}")
      * @var string
      */
     protected $balancer = 'random';
@@ -94,7 +94,7 @@ class PropertyPoolConfig extends PoolProperties
     /**
      * the default provider is consul provider
      *
-     * @Value(name="${config.test.test.provider}")
+     * @Value(name="${config.test.test.provider}", env="${TEST_PROVIDER}")
      * @var string
      */
     protected $provider = 'consul';
