+++ test/.env.example
@@ -118,3 +118,20 @@ CONSUL_REGISTER_CHECK_NAME=swoft
 CONSUL_REGISTER_CHECK_TCP=127.0.0.1:8099
 CONSUL_REGISTER_CHECK_INTERVAL=10
 CONSUL_REGISTER_CHECK_TIMEOUT=1
+
+# test config
+TEST_NAME=test
+TEST_URI=127.0.0.1:6378
+TEST_MAX_IDEL=2
+TEST_MAX_ACTIVE=2
+TEST_MAX_WAIT=2
+TEST_TIMEOUT=2
+TEST_BALANCER=r1
+TEST_USE_PROVIDER=true
+TEST_PROVIDER=c1
+
+# consul provider
+PROVIDER_CONSUL_ADDRESS=http://127.0.0.1:82
+PROVIDER_CONSUL_TAGS=1,2
+PROVIDER_CONSUL_TIMEOUT=2
+PROVIDER_CONSUL_INTERVAL=2
