+++ test/config/properties/components.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: limx
+ * Date: 2018/12/9
+ * Time: 10:04 AM
+ */
+return [
+    'custom' => [
+        'SwoftTest',
+        'SwoftTest\\Testing\\Bean' => BASE_PATH . '/Testing/Bean',
+        'SwoftTest\\Testing\\Bean2' => '@root/Testing/Bean2',
+    ]
+];
